fix(PlotOutputFn): using PlotType == PDF caused empty plots
ireader and jreader were not found because they came to late in the
code after the pdf name was specified. blank plots were being created
because the ageing_comparison was returning return() if there were
no double reads. using a return of the data frame and a check for
NULL fixes this and the pdf is deleted if it is empty. there are many
more things that could be fixed here, but lack of time, leads to hacks
rather than complete fixes. such as dev.off() should be called only if
a new plot is created rather than all times. the if (PlotType) is not the
best and the type of file should be able to be sent to ageing_comparison
instead of wrapping outside of it.

Thanks to Paul Spencer for warning us of these errors and for helping to
suggest a fix for them!
diff --git a/R/PlotOutputFn.R b/R/PlotOutputFn.R
index e8a78d2..6ff9a98 100644
--- a/R/PlotOutputFn.R
+++ b/R/PlotOutputFn.R
@@ -108,18 +108,20 @@ PlotOutputFn <-

   ####################################################################
   # Plot comparison of data for each pair of readers
-
+browser()
   if(1 %in% subplot){
-    if(PlotType == "PDF"){
-      pdf(paste0(ReaderNames[ireader],
-                 " vs ", ReaderNames[jreader], ".pdf"),
-          width = 6, height = 6)
-    }

     # make plots of input data for each reader pair
     for(ireader in 1:(Nreaders-1)){
       for(jreader in (ireader+1):Nreaders){
-        ageing_comparison(xvec = DataExpanded[,ireader],
+        if (PlotType == "PDF") {
+          pdfname <- paste0(ReaderNames[ireader],
+            " vs ", ReaderNames[jreader], ".pdf")
+          pdf(pdfname,
+            width = 6, height = 6)
+        }
+
+        out <- ageing_comparison(xvec = DataExpanded[,ireader],
                           yvec = DataExpanded[,jreader],
                           xlab = ReaderNames[ireader],
                           ylab = ReaderNames[jreader],
@@ -131,6 +133,12 @@ PlotOutputFn <-
                               " vs ", ReaderNames[jreader], ".png"),
                           verbose = FALSE,
                           ...)
+        if (PlotType == "PDF") {
+          dev.off()
+          if (is.null(out)) {
+            unlink(pdfname)
+          }
+        }
       }
     }
   } # end check for whether subplot 1 was requested
diff --git a/R/ageing_comparison.R b/R/ageing_comparison.R
index ff307d6..1196ab1 100644
--- a/R/ageing_comparison.R
+++ b/R/ageing_comparison.R
@@ -104,4 +104,5 @@ ageing_comparison <- function(xvec, yvec, scale.pts=2,
   if(png){
     dev.off()
   }
+  invisible(df1)
 }